+++ deploy/python/keypoint_postprocess.py
@@ -356,7 +356,7 @@ def affine_transform(pt, t):
     return new_pt[:2]
 
 
-def affine_backto_orgimages(keypoint_result, batch_records):
+def translate_to_ori_images(keypoint_result, batch_records):
     kpts, scores = keypoint_result['keypoint']
     kpts[..., 0] += batch_records[:, 0:1]
     kpts[..., 1] += batch_records[:, 1:2]
