+++ deploy/python/mot_sde_infer.py
@@ -19,7 +19,7 @@ import cv2
 import numpy as np
 import paddle
 from benchmark_utils import PaddleInferBenchmark
-from preprocess import preprocess, NormalizeImage, Permute, LetterBoxResize
+from preprocess import preprocess
 from tracker import DeepSORTTracker
 from ppdet.modeling.mot import visualization as mot_vis
 from ppdet.modeling.mot.utils import Timer as MOTTimer
@@ -29,7 +29,8 @@ from paddle.inference import Config
 from paddle.inference import create_predictor
 from utils import argsparser, Timer, get_current_memory_mb
 from infer import get_test_images, print_arguments, PredictConfig, Detector
-from mot_infer import create_inputs, load_predictor, write_mot_results
+from mot_jde_infer import write_mot_results
+from infer import load_predictor
 
 # Global dictionary
 MOT_SUPPORT_MODELS = {'DeepSORT'}
@@ -73,23 +74,6 @@ def clip_box(xyxy, input_shape, im_shape, scale_factor):
     return xyxy
 
 
-def get_crops(xyxy, ori_img, pred_scores, w, h):
-    crops = []
-    keep_scores = []
-    xyxy = xyxy.astype(np.int64)
-    ori_img = ori_img.transpose(1, 0, 2)  # [h,w,3]->[w,h,3]
-    for i, bbox in enumerate(xyxy):
-        if bbox[2] <= bbox[0] or bbox[3] <= bbox[1]:
-            continue
-        crop = ori_img[bbox[0]:bbox[2], bbox[1]:bbox[3], :]
-        crops.append(crop)
-        keep_scores.append(pred_scores[i])
-    if len(crops) == 0:
-        return [], []
-    crops = preprocess_reid(crops, w, h)
-    return crops, keep_scores
-
-
 def preprocess_reid(imgs,
                     w=64,
                     h=192,
@@ -109,7 +93,7 @@ def preprocess_reid(imgs,
     return im_batch
 
 
-class MOT_Detector(object):
+class SDE_Detector(Detector):
     """
     Args:
         pred_config (object): config of model, defined by `Config(model_dir)`
@@ -130,37 +114,26 @@ class MOT_Detector(object):
                  model_dir,
                  device='CPU',
                  run_mode='fluid',
+                 batch_size=1,
                  trt_min_shape=1,
                  trt_max_shape=1088,
                  trt_opt_shape=608,
                  trt_calib_mode=False,
                  cpu_threads=1,
                  enable_mkldnn=False):
-        self.pred_config = pred_config
-        self.predictor, self.config = load_predictor(
-            model_dir,
-            run_mode=run_mode,
+        super(SDE_Detector, self).__init__(
+            pred_config=pred_config,
+            model_dir=model_dir,
             device=device,
-            min_subgraph_size=self.pred_config.min_subgraph_size,
-            use_dynamic_shape=self.pred_config.use_dynamic_shape,
+            run_mode=run_mode,
+            batch_size=batch_size,
             trt_min_shape=trt_min_shape,
             trt_max_shape=trt_max_shape,
             trt_opt_shape=trt_opt_shape,
             trt_calib_mode=trt_calib_mode,
             cpu_threads=cpu_threads,
             enable_mkldnn=enable_mkldnn)
-        self.det_times = Timer()
-        self.cpu_mem, self.gpu_mem, self.gpu_util = 0, 0, 0
-
-    def preprocess(self, im):
-        preprocess_ops = []
-        for op_info in self.pred_config.preprocess_infos:
-            new_op_info = op_info.copy()
-            op_type = new_op_info.pop('type')
-            preprocess_ops.append(eval(op_type)(**new_op_info))
-        im, im_info = preprocess(im, preprocess_ops)
-        inputs = create_inputs(im, im_info)
-        return inputs
+        assert batch_size == 1, "The JDE Detector only supports batch size=1 now"
 
     def postprocess(self, boxes, input_shape, im_shape, scale_factor,
                     threshold):
@@ -214,7 +187,7 @@ class MOT_Detector(object):
         return pred_bboxes, pred_scores
 
 
-class MOT_ReID(object):
+class SDE_ReID(object):
     def __init__(self,
                  pred_config,
                  model_dir,
@@ -300,6 +273,24 @@ class MOT_ReID(object):
         self.det_times.img_num += 1
         return online_tlwhs, online_scores, online_ids
 
+    def get_crops(self, xyxy, ori_img, pred_scores, w, h):
+        self.det_times.preprocess_time_s.start()
+        crops = []
+        keep_scores = []
+        xyxy = xyxy.astype(np.int64)
+        ori_img = ori_img.transpose(1, 0, 2)  # [h,w,3]->[w,h,3]
+        for i, bbox in enumerate(xyxy):
+            if bbox[2] <= bbox[0] or bbox[3] <= bbox[1]:
+                continue
+            crop = ori_img[bbox[0]:bbox[2], bbox[1]:bbox[3], :]
+            crops.append(crop)
+            keep_scores.append(pred_scores[i])
+        if len(crops) == 0:
+            return [], []
+        crops = preprocess_reid(crops, w, h)
+        self.det_times.preprocess_time_s.end()
+        return crops, keep_scores
+
 
 def predict_image(detector, reid_model, image_list):
     results = []
@@ -307,21 +298,22 @@ def predict_image(detector, reid_model, image_list):
         frame = cv2.imread(img_file)
         if FLAGS.run_benchmark:
             pred_bboxes, pred_scores = detector.predict(
-                frame, FLAGS.threshold, warmup=10, repeats=10)
+                [frame], FLAGS.threshold, warmup=10, repeats=10)
             cm, gm, gu = get_current_memory_mb()
             detector.cpu_mem += cm
             detector.gpu_mem += gm
             detector.gpu_util += gu
             print('Test iter {}, file name:{}'.format(i, img_file))
         else:
-            pred_bboxes, pred_scores = detector.predict(frame, FLAGS.threshold)
+            pred_bboxes, pred_scores = detector.predict([frame],
+                                                        FLAGS.threshold)
 
         # process
         bbox_tlwh = np.concatenate(
             (pred_bboxes[:, 0:2],
              pred_bboxes[:, 2:4] - pred_bboxes[:, 0:2] + 1),
             axis=1)
-        crops, pred_scores = get_crops(
+        crops, pred_scores = reid_model.get_crops(
             pred_bboxes, frame, pred_scores, w=64, h=192)
 
         if FLAGS.run_benchmark:
@@ -330,14 +322,16 @@ def predict_image(detector, reid_model, image_list):
         else:
             online_tlwhs, online_scores, online_ids = reid_model.predict(
                 crops, bbox_tlwh, pred_scores)
-
             online_im = mot_vis.plot_tracking(
                 frame, online_tlwhs, online_ids, online_scores, frame_id=i)
 
             if FLAGS.save_images:
                 if not os.path.exists(FLAGS.output_dir):
                     os.makedirs(FLAGS.output_dir)
-                cv2.imwrite(os.path.join(FLAGS.output_dir, img_file), online_im)
+                img_name = os.path.split(img_file)[-1]
+                out_path = os.path.join(FLAGS.output_dir, img_name)
+                cv2.imwrite(out_path, online_im)
+                print("save result to: " + out_path)
 
 
 def predict_video(detector, reid_model, camera_id):
@@ -367,14 +361,13 @@ def predict_video(detector, reid_model, camera_id):
         if not ret:
             break
         timer.tic()
-        pred_bboxes, pred_scores = detector.predict(frame, FLAGS.threshold)
+        pred_bboxes, pred_scores = detector.predict([frame], FLAGS.threshold)
         timer.toc()
-
         bbox_tlwh = np.concatenate(
             (pred_bboxes[:, 0:2],
              pred_bboxes[:, 2:4] - pred_bboxes[:, 0:2] + 1),
             axis=1)
-        crops, pred_scores = get_crops(
+        crops, pred_scores = reid_model.get_crops(
             pred_bboxes, frame, pred_scores, w=64, h=192)
 
         online_tlwhs, online_scores, online_ids = reid_model.predict(
@@ -413,7 +406,7 @@ def predict_video(detector, reid_model, camera_id):
 
 def main():
     pred_config = PredictConfig(FLAGS.model_dir)
-    detector = MOT_Detector(
+    detector = SDE_Detector(
         pred_config,
         FLAGS.model_dir,
         device=FLAGS.device,
@@ -426,7 +419,7 @@ def main():
         enable_mkldnn=FLAGS.enable_mkldnn)
 
     pred_config = PredictConfig(FLAGS.reid_model_dir)
-    reid_model = MOT_ReID(
+    reid_model = SDE_ReID(
         pred_config,
         FLAGS.reid_model_dir,
         device=FLAGS.device,
